pyproject.toml
[project]
name = "wow-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
]
